airflow/requirements.txt
apache-airflow
pandas
requests
psycopg2-binary
sqlalchemy
python-dotenv